# Small performance improvement.