@samuelsonric 2025-02-11
I just realized that you could supply a `scratch` vector to `sort!`.

Now we're at
```julia
BenchmarkTools.Trial: 46 samples with 1 evaluation per sample.
 Range (min … max):  104.671 ms … 118.951 ms  ┊ GC (min … max): 0.00% … 4.10%
 Time  (median):     107.956 ms               ┊ GC (median):    2.81%
 Time  (mean ± σ):   109.165 ms ±   3.183 ms  ┊ GC (mean ± σ):  2.85% ± 1.33%

         ▁▁▁█▄▄██ ▁           ▁                                  
  ▆▁▁▁▆▁▆████████▁█▆▆▁▁▆▆▆▆▁▁▆█▆▁▁▆▁▁▁▆▁▆▁▁▁▁▁▁▁▁▁▁▆▆▁▆▁▁▁▁▁▁▁▆ ▁
  105 ms           Histogram: frequency by time          119 ms <

 Memory estimate: 218.91 MiB, allocs estimate: 17.
```